Fix a very subtle bug in shaders that have params whose default values are expressions that need to be evaluated at runtime ("init ops") -- because we ran the init ops when binding, the lazy execution of layers was setting the uniform/varyingness of the downstream connection BEFORE the upstream value was assigned (which could change it from uniform to varying).
The solution is that the init ops should be executed as part of the
RUNNING of the shader, not the BINDING.  Then it always happens in the
correct order, even in the presence of lazy layers. As a bonus, this
also is a performance improvement, since the same bug would trigger
evaluation of earlier layers in cases where it wasn't needed.

http://codereview.appspot.com/195055/show
Review by: Chris Kulla