Bind parameters individually as they are needed (as revealed by the 'useparam' instruction), which is even more efficient than binding when the shader is run, since some params (and hence their init ops or geom lookups) won't ever be used, depending on the code path taken.  This seems to speed up our production scenes by about 10%.  This optimization was suggested by Chris Kulla.
Along the way, we fix two pre-existing issues with useparam:

1. We weren't doing proper useparam for output params; now we useparam
them all up front, which is necessary in case the output params aren't
directly referenced in the main body of the shader.

2. We were issuing useparam instructions every time we access a param.
But now I have it rigged so that once it does a 'useparam'
unconditionally (i.e. in the main shader body, and not inside an 'if' or
loop), it won't useparam the same parameter again.  This eliminates a
lot of redundant useparam instructions.

http://codereview.appspot.com/217043/show
Review by: ckulla