Add key attributes to ClosureComponent (issue#2216048)
This makes it possible to add render specific args
to closures. It works like keyword arguments in the
constructor, and in order for them to get passed
they have to be registered by the renderer in
register_closure. See the new macros in genclosure.h

Also closures are now printed with a single function,
print_on virtual method is no longer used.